+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-22 21:58:18
-// Commit: fix(core): fix memory leak by cancelling core subscription
+// Updated at: 2026-08-23 09:31:54
+// Commit: fix(core): correct layout constraints and overflow in core
 
 class coreComponentState {
-  final String commitRef = '3818';
-  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
-  final DateTime timestamp = DateTime.parse('2026-08-22T21:58:18+05:00');
+  final String commitRef = '3819';
+  final String lastMessage = 'fix(core): correct layout constraints and overflow in core';
+  final DateTime timestamp = DateTime.parse('2026-08-23T09:31:54+05:00');
 }