+++ CHANGELOG.md
@@ -3820,3 +3820,4 @@
 - [2026-08-22 20:54] chore(security): update dependencies and cleanup unused imports in security
 - [2026-08-22 21:27] chore(cache): update dependencies and cleanup unused imports in cache
 - [2026-08-22 21:58] fix(core): fix memory leak by cancelling core subscription
+- [2026-08-23 09:31] fix(core): correct layout constraints and overflow in core